+++ lib/src/extensions/flutter_extensions/platform_extension.dart
@@ -70,6 +70,66 @@ extension FHUPlatformExtensionNullable on BuildContext? {
   bool get isAppleWeb => targetPlatform.isAppleWeb;
 }
 
+/// Extension on nullable [ThemeData] to provide various platform checks.
+extension FHUThemeDataPlatformExtensionNullable on ThemeData? {
+  /// Gets the [TargetPlatform] of the nullable [ThemeData].
+  TargetPlatform? get platform => this?.platform;
+
+  /// Checks if the platform is a mobile device (iOS or Android) natively.
+  bool get isMobile => platform.isMobile;
+
+  /// Checks if the platform is iOS natively.
+  bool get isIOS => platform.isIOS;
+
+  /// Checks if the platform is Android natively.
+  bool get isAndroid => platform.isAndroid;
+
+  /// Checks if the platform is a desktop OS (Linux, macOS, Windows) natively.
+  bool get isDesktop => platform.isDesktop;
+
+  /// Checks if the platform is macOS natively.
+  bool get isMacOS => platform.isMacOS;
+
+  /// Checks if the platform is Windows natively.
+  bool get isWindows => platform.isWindows;
+
+  /// Checks if the platform is Linux natively.
+  bool get isLinux => platform.isLinux;
+
+  /// Checks if the platform is Fuchsia natively.
+  bool get isFuchsia => platform.isFuchsia;
+
+  /// Checks if the platform is an Apple platform (macOS or iOS) natively.
+  bool get isApple => platform.isApple;
+
+  /// Checks if the platform is a mobile browser (iOS or Android).
+  bool get isMobileWeb => platform.isMobileWeb;
+
+  /// Checks if the platform is iOS in a web browser.
+  bool get isIOSWeb => platform.isIOSWeb;
+
+  /// Checks if the platform is Android in a web browser.
+  bool get isAndroidWeb => platform.isAndroidWeb;
+
+  /// Checks if the platform is a desktop browser (Linux, macOS, Windows).
+  bool get isDesktopWeb => platform.isDesktopWeb;
+
+  /// Checks if the platform is macOS in a web browser.
+  bool get isMacOsWeb => platform.isMacOsWeb;
+
+  /// Checks if the platform is Windows in a web browser.
+  bool get isWindowsWeb => platform.isWindowsWeb;
+
+  /// Checks if the platform is Linux in a web browser.
+  bool get isLinuxWeb => platform.isLinuxWeb;
+
+  /// Checks if the platform is Fuchsia in a web browser.
+  bool get isFuchsiaWeb => platform.isFuchsiaWeb;
+
+  /// Checks if the platform is an Apple platform (macOS or iOS) in a web browser.
+  bool get isAppleWeb => platform.isAppleWeb;
+}
+
 /// Extension on [TargetPlatform] to provide various platform checks.
 extension FHUTargetPlatformExtension on TargetPlatform? {
   /// Checks if the platform is a mobile device (iOS or Android) natively.
@@ -141,6 +201,16 @@ extension FHUTargetPlatformExtension on TargetPlatform? {
 
 /// A platform-independent class providing information about the operating system,
 /// with web compatibility considerations.
+///
+/// Similar to the `Platform` class from the `dart:io` library, but this one
+/// respects the web environment.
+/// A platform-independent class providing information about the operating system,
+/// with special considerations for web compatibility.
+///
+/// This class serves as a unified interface for interacting with platform-specific
+/// details, abstracting away the differences between native and web environments.
+/// It mirrors the functionality of the `Platform` class from the `dart:io` library,
+/// but ensures consistent behavior even when running Dart code in a web browser.
 abstract class PlatformEnv {
   /// Returns the `TargetPlatform` corresponding to the current operating system.
   ///