+++ lib/src/value_notifier/widgets/widgets.dart
@@ -1,2 +1,1 @@
-export 'box.dart';
 export 'listenables_builder.dart';